src/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
Flask-Migrate==4.0.5
alembic==1.13.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
PyMySQL==1.1.0
